chore: remove unnecessary license clause
Setting this field makes the entire GPLv2 license text appear when
viewing the package metadata e.g with `pip show flake8-builtins`.

I feel this is a bit too verbose, and unnecessary.

As described in the [python packaging guide] it is not necessary to
specify this field when using a standard license. Actually this might be
more confusing to users, as they may avoid the package if a modified
license is implied:

> If you are using a standard, well-known license, it is not necessary
> to use this field. Instead, you should one of the classifiers starting
> with License ::. (As a general rule, it is a good idea to use a
> standard, well-known license, both to avoid confusion and because some
> organizations avoid software whose license is unapproved.)

[python packaging guide]: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license