Feat: add initial support for `TypedDict`
`TypedDicts` are used only within the type system for analysis and
represent the keys of a dictionary and their corresponding types.
They are a useful tool when providing type hints of a JSON payload or
similar.

For example the following code snippet will raise a false-positive alert
from this flake8 plug-in:
```python
class ApiPage(TypedDict):
  url: str
  next: str
  data: list[dict[str, Any]]
```

This is because a `TypedDict` is analysed as if it was a standard class.
Since it is only used in the type system it is safe to ignore anything
that inherits from `TypedDict` in this plugin.